scripts/twitter-poem-bot/requirements.txt
google-generativeai>=0.3.0
tweepy>=4.14.0
